requirement.txt
torch
transformers
tokenizers
pandas
pyarrow
pillow
